+++ resources/views/components/dropdown-title.blade.php
@@ -7,5 +7,5 @@
 		class="avatar avatar-small"
 		src="{{ Storage::url(Auth::user()->avatar ?? '/avatars/default-avatar.png') }}"
 		alt="Avatar">
-	<span>{{ Auth::user()->display_name }}</span>
+	<span>{{ Auth::user()->display_name ?? 'Joe Shmoe' }}</span>
 </button>